implement cooked physx mesh caching + documentation + other improvements
