+++ Source/CesiumRuntime/CesiumRuntime.Build.cs
@@ -135,7 +135,7 @@ public class CesiumRuntime : ModuleRules
         PublicDependencyModuleNames.AddRange(
             new string[]
             {
-                "Core",
+                "Core"
                 // ... add other public dependencies that you statically link with here ...
             }
         );
